Separate API modality contracts and share multipart field validation
